requirements.txt
fastapi[standard]
yfinance
pandas
streamlit
requests
plotly
langchain
langchain-community
langchain-core
langchain-text-splitters
chromadb
langchain-chroma
sentence-transformers
redis
langchain-ollama
langchain-huggingface
langchain-google-genai
langchain-anthropic
python-dotenv
sqlalchemy
psycopg2-binary
uvicorn
pandas_ta
